Made simple blockchain, changed package name for search engine
