+++ java-blockchain/src/main/java/com/nbugaenco/blockchain/util/HashUtils.java
@@ -1,0 +1,29 @@
+package com.nbugaenco.blockchain.util;
+
+import java.nio.charset.StandardCharsets;
+import java.security.MessageDigest;
+
+public class HashUtils {
+
+    private HashUtils() {
+        throw new IllegalStateException("Utility class");
+    }
+
+    /* Applies Sha256 to a string and returns a hash. */
+    public static String applySha256(String input) {
+        try {
+            MessageDigest digest = MessageDigest.getInstance("SHA-256");
+            /* Applies sha256 to our input */
+            byte[] hash = digest.digest(input.getBytes(StandardCharsets.UTF_8));
+            StringBuilder hexString = new StringBuilder();
+            for (byte elem : hash) {
+                String hex = Integer.toHexString(0xff & elem);
+                if (hex.length() == 1) hexString.append('0');
+                hexString.append(hex);
+            }
+            return hexString.toString();
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+}
