+++ simple-search-engine-maven/src/main/java/com/nbugaenco/searchengine/model/SearchStrategy.java
@@ -1,0 +1,7 @@
+package com.nbugaenco.searchengine.model;
+
+public enum SearchStrategy {
+    ALL,
+    ANY,
+    NONE
+}
