--- simple-search-engine-maven/src/main/java/searchengine/model/SearchStrategy.java
@@ -1,7 +1,0 @@
-package searchengine.model;
-
-public enum SearchStrategy {
-    ALL,
-    ANY,
-    NONE
-}
