+++ simple-search-engine-maven/src/test/java/com/nbugaenco/searchengine/util/SearchEngineUtilsTest.java
@@ -1,0 +1,17 @@
+package com.nbugaenco.searchengine.util;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class SearchEngineUtilsTest {
+
+    @Test
+    @DisplayName("Test invalid CLI arguments")
+    void readFileWithWrongArgs() {
+        IllegalArgumentException ex = Assertions.assertThrows(IllegalArgumentException.class,
+                () -> SearchEngineUtils.readFile(new String[]{"--daa", "text.txt"}));
+
+        Assertions.assertTrue(ex.getMessage().contains("Invalid argument"));
+    }
+}
